feat: Add main page listing(infinite scroll, search, sorting) , Fix location type, post response dto, yml(sql) (#41)
- 메인 페이지 리스팅
- 무한스크롤을 고려한 페이지네이션 (슬라이스)
- 검색 기능 추가
- 정렬 기능 추가
- location 타입 double로 수정
- 메인 페이지 리스팅에 썸네일만 보내기 위해 postResponseDto 수정 (조회수, 좋아요, 닉네임 null 로 나오던 것도 수정)
- yml 파일에 SQL 관련 추가 (오히려 지저분할 수도 있으니 필요없으면 지우세요)

Resolves #29 #33 #34 #38

+++ yoseobara/src/main/java/com/final2/yoseobara/dto/request/PostRequestDto.java
@@ -13,6 +13,6 @@ public class PostRequestDto {
     private String title;
     private String content;
     private String address;
-    private HashMap<String,Float> location;
+    private HashMap<String,Double> location;
 
 }
